third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/animation-delay-expected.txt

This is a testharness.js-based test.
FAIL Can set 'animation-delay' to CSS-wide keywords: initial assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'animation-delay' to CSS-wide keywords: inherit assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'animation-delay' to CSS-wide keywords: unset assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'animation-delay' to CSS-wide keywords: revert assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'animation-delay' to var() references:  var(--A) assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
FAIL Can set 'animation-delay' to a time: 0s Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'animation-delay' to a time: -3.14ms Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'animation-delay' to a time: 3.14s Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'animation-delay' to a time: calc(0s + 0ms) Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting 'animation-delay' to a length throws TypeError
PASS Setting 'animation-delay' to a percent throws TypeError
PASS Setting 'animation-delay' to an angle throws TypeError
PASS Setting 'animation-delay' to a flexible length throws TypeError
PASS Setting 'animation-delay' to a number throws TypeError
PASS Setting 'animation-delay' to a URL throws TypeError
PASS Setting 'animation-delay' to a transform throws TypeError
Harness: the test ran to completion.
